MinGW compatibility.
git-svn-id: http://www.bci2000.org/svn/trunk@1077 7ef59cdf-0624-0410-a474-e34ce222aafa

+++ src/shared/UBCI2000Data.h
@@ -59,7 +59,7 @@ class BCI2000DATA
   int                InitializeTotal( const char* new_filename, int buffer_size  = defaultBufSize );
 
   // Please, use GetSamplingRate() instead.
-  int                GetSampleFrequency() const   { return mSamplingRate; }
+  int                GetSampleFrequency() const   { return static_cast<int>( mSamplingRate ); }
 
  private:
   void               Reset();
@@ -84,7 +84,7 @@ class BCI2000DATA
   std::ifstream      mFile;
   std::string        mFilename,
                      mFileFormatVersion;
-                     
+
   SignalType         mSignalType;
   int                mDataSize;
   GenericSignal::value_type ( *mfpReadValueBinary )( const char* );
